Update Strong Tag Style & Move Code About Utility Style Map Code
