+++ tailwind.config.js
@@ -8,7 +8,6 @@ module.exports = {
         "font-white": "#dadada",
         "font-gray": "#a2aab3",
         "font-quote": "#3eb4c0",
-        bold: "#ff82b2",
         link: "#79a9ed",
         "link-hover": "#adc9f2",
         "background-dark": "#1c2127",
